Refactor: Refactor how Tribble 4-way races are managed.
After staring at code for a long time yesterday, I realized that `getInnermostClient()` is strictly used to deal with Tribble 4-way races, and there is always an associated `writeDescriptor()` nearby. Life gets easier if we combine these two concepts into one call.

In the process, we also get closer to achieving e-order for three-party handoffs. However, embargoes are not yet implemeted for 3PH, so this doesn't actually help yet.