3PCF: Fix bug with embargos, forwarding, and disconnects.
If a call had been forwarded along a connection that was disconnected before the embargo was lifted, previously the embargo would erroneously be lifted. Now it'll fail.

This maintains e-order guarantees: If you make several calls in a row, and one call cannot be delivered for some reason, later calls must also fail.